Refactor endpoint registration
This commit moves the endpoint registration logic to its own file.
The main changes are that the recursive logic for retries was replaced
by an iteration. Each expected errors is thrown with a more detailed
cause.

For example, for rate limit errors, an `EndpointRegistrationError`
is thrown. It has a `cause` property with the `name`
`"LoginRateLimitExceeded"` and a `data` property with the list of
performed HTTP exchanges.

Closes ocilo/skype-http#55
